% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_K_factor.R
\name{get_K_factor}
\alias{get_K_factor}
\title{Get information on K Factors of PNCT count sites}
\usage{
get_K_factor(id, year)
}
\arguments{
\item{id}{PNCT equipment id (given by get_PNCT_sites()).}

\item{year}{traffic data for a given year. Input one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}
}
\value{
Returns a data frame with K factors data for a given count location
}
\description{
Gathers information on K Factors (in portuguese, Fatores K) for a given count location using the DNIT Web API.
}
\examples{
temp1 <- get_K_factor(id = 414,year=2018)
}
