% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Equipments.R
\name{get_PNCT_sites}
\alias{get_PNCT_sites}
\title{Get information on PNCT count locations}
\usage{
get_PNCT_sites(year = "all", geo = FALSE)
}
\arguments{
\item{year}{equipments running on a given year. Input "all" for all available equipments (default), or one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}

\item{geo}{if TRUE, returns a simple features (sf) object using geographical coordinates from count locations. Defaults FALSE.}
}
\value{
Returns a data frame or sf object regarding PNCT count locations
}
\description{
Gathers information on PNCT count locations using the DNIT Web API, for a given year or all available years. Returns data on data frame or sf format.
}
\examples{
temp1 <- get_PNCT_sites(year="all",geo=T)
temp2 <- get_PNCT_sites(year=2014,geo=F)
}
