% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_factors.R
\name{get_seasonal_factor}
\alias{get_seasonal_factor}
\title{Get information on Seasonal Factors of PNCT count sites}
\usage{
get_seasonal_factor(id, year, type = c("month", "week", "weekday"))
}
\arguments{
\item{id}{PNCT equipment id (given by get_PNCT_sites()).}

\item{year}{traffic data for a given year. Input one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}

\item{type}{type of seasonal factor: "month" for Monthly factors, "week" for Weekly factors, or "weekday" for seasonal factors among days of the week.}
}
\value{
Returns a data frame with seasonal factors data for a given count location
}
\description{
Gathers information on Seasonal Factors (in portuguese, Fatores de Sazonalidade) for a given count location using the DNIT Web API.
}
\examples{
temp1 <- get_seasonal_factors(id = 414,year=2018,type="month")
}
