% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VDMD.R
\name{get_VDMD}
\alias{get_VDMD}
\title{Get information on Daily Average Daily Traffic (DADT) of PNCT count sites}
\usage{
get_VDMD(id, year)
}
\arguments{
\item{id}{PNCT equipment id (given by get_PNCT_sites()).}

\item{year}{traffic data for a given year. Input one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}
}
\value{
Returns a data frame with WADT data for a given count location
}
\description{
Gathers information on DADT (in portuguese, VDMd, Volume Diário Médio Diário) for a given count location using the DNIT Web API.
}
\examples{
temp1 <- get_VDMS(id = 414,year="all")
temp2 <- get_VDMS(id = 414,year=2018)
}
