% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VDMM.R
\name{get_VDMM}
\alias{get_VDMM}
\title{Get information on Monthly Average Daily Traffic (MADT) of PNCT count sites}
\usage{
get_VDMM(id, year = "all")
}
\arguments{
\item{id}{PNCT equipment id (given by get_PNCT_sites()).}

\item{year}{traffic data for a given year. Input "all" for all available data (default), or one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}
}
\value{
Returns a data frame with MADT data for a given count location
}
\description{
Gathers information on MADT (in portuguese, VDMm, Volume Diário Médio Mensal) for a given count location using the DNIT Web API.
}
\examples{
temp1 <- get_VDMM(id = 414,year="all")
temp1 <- get_VDMM(id = 414,year=2018)
}
