% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VH.R
\name{get_VH}
\alias{get_VH}
\title{Get information on Hourly Traffic Volumes of PNCT count sites}
\usage{
get_VH(id, year, month, day)
}
\arguments{
\item{id}{PNCT equipment id (given by get_PNCT_sites()).}

\item{year}{traffic data for a given year. Input one of the following years: 2014,2015,2016,2017,2018,2019 or 2021.}

\item{month}{traffic data for a given month of the year. One of 1,2,3,4,5,6,7,8,9,10,11,12.}

\item{day}{traffic data for a given day of the month. Input should be between 1 - 31.}
}
\value{
Returns a data frame with traffic volume data for a given count location
}
\description{
Gathers information on Hourly Traffic Volumes (in portuguese, Volumes Horários) for a given count location using the DNIT Web API.
}
\examples{
temp1 <- get_VH(id = 414,year=2018,month=1,day=1)
}
